refactor(archive): move the activity wire types to dango-archive-types
Move `Page` / `PageInfo` (from the httpd) and `Event` / `Transaction` /
`AddressRole` / `UnitKind` / `EventType` (from the projection) into the
shared `dango-archive-types` crate, adding `Deserialize`, so clients can
consume the exact structs the read API serializes and the wire format
cannot drift. The utoipa schema derives ride behind a new `utoipa`
feature, enabled by the httpd and the projection, so clients never pull
utoipa.

The projection keeps its storage glue (row structs, stored-code and
stored-bytes decoders, hydration); `TryFrom<transactions::Model>`
becomes the free function `transaction_from_model`, since the impl
would be an orphan with `Transaction` now foreign. The httpd re-exports
`Page` / `PageInfo`, so downstream imports are unchanged.

No behavior change; the wire format is pinned by new serde round-trip
tests and the untouched archive e2e suite.

---------
