+++ dango/archive/projection/src/activity/event_type.rs
@@ -1,95 +1,10 @@
-use {
-    dango_primitives::{Addr, FlatEvent},
-    serde::{Deserialize, Serialize},
-};
-
-/// Compact discriminant for a [`FlatEvent`], stored in `activity_events.event_type`
-/// and used to key the projection's priority / involvement configuration.
-///
-/// Mirrors `FlatEvent`'s variants one-to-one. The `#[repr(i16)]` values are the
-/// **stored codes**: they are part of the on-disk schema, so existing codes must
-/// never be renumbered (append-only). [`From<&FlatEvent>`] is exhaustive, so a
-/// new upstream variant is a compile error here until it is given a code.
-///
-/// Serialized snake_case (`transfer`, `contract_event`, …): the spelling the
-/// read API accepts to filter the activity feeds by type and surfaces as an
-/// event's `type`.
-#[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq, Eq, Hash, utoipa::ToSchema)]
-#[serde(rename_all = "snake_case")]
-#[repr(i16)]
-pub enum EventType {
-    Configure = 0,
-    Upgrade = 1,
-    Transfer = 2,
-    Upload = 3,
-    Instantiate = 4,
-    Execute = 5,
-    Migrate = 6,
-    Reply = 7,
-    Authenticate = 8,
-    Backrun = 9,
-    Withhold = 10,
-    Finalize = 11,
-    Cron = 12,
-    Guest = 13,
-    ContractEvent = 14,
-}
-
-impl EventType {
-    /// The stored discriminant (the value written to `event_type`).
-    #[must_use]
-    pub fn code(self) -> i16 {
-        self as i16
-    }
-
-    /// The variant for a stored discriminant, or `None` if `code` is unknown —
-    /// the inverse of [`code`](Self::code), used to surface `event_type` in the
-    /// read API. Spelled out (not `transmute`) so an out-of-range value read
-    /// from the database is a clean `None`, never undefined behaviour.
-    #[must_use]
-    pub fn from_code(code: i16) -> Option<Self> {
-        Some(match code {
-            0 => Self::Configure,
-            1 => Self::Upgrade,
-            2 => Self::Transfer,
-            3 => Self::Upload,
-            4 => Self::Instantiate,
-            5 => Self::Execute,
-            6 => Self::Migrate,
-            7 => Self::Reply,
-            8 => Self::Authenticate,
-            9 => Self::Backrun,
-            10 => Self::Withhold,
-            11 => Self::Finalize,
-            12 => Self::Cron,
-            13 => Self::Guest,
-            14 => Self::ContractEvent,
-            _ => return None,
-        })
-    }
-}
-
-impl From<&FlatEvent> for EventType {
-    fn from(event: &FlatEvent) -> Self {
-        match event {
-            FlatEvent::Configure(_) => Self::Configure,
-            FlatEvent::Upgrade(_) => Self::Upgrade,
-            FlatEvent::Transfer(_) => Self::Transfer,
-            FlatEvent::Upload(_) => Self::Upload,
-            FlatEvent::Instantiate(_) => Self::Instantiate,
-            FlatEvent::Execute(_) => Self::Execute,
-            FlatEvent::Migrate(_) => Self::Migrate,
-            FlatEvent::Reply(_) => Self::Reply,
-            FlatEvent::Authenticate(_) => Self::Authenticate,
-            FlatEvent::Backrun(_) => Self::Backrun,
-            FlatEvent::Withhold(_) => Self::Withhold,
-            FlatEvent::Finalize(_) => Self::Finalize,
-            FlatEvent::Cron(_) => Self::Cron,
-            FlatEvent::Guest(_) => Self::Guest,
-            FlatEvent::ContractEvent(_) => Self::ContractEvent,
-        }
-    }
-}
+/// The compact [`FlatEvent`] discriminant, moved to [`dango_archive_types`] so
+/// clients (e.g. the SDK's `ArchiveClient`) share the exact spellings and
+/// stored codes; re-exported here for the projection's many internal users.
+/// What stays below is the projection-only taxonomy: which columns a
+/// `FlatEvent` populates.
+pub use dango_archive_types::EventType;
+use dango_primitives::{Addr, FlatEvent};
 
 // ---- per-event taxonomy used by the `events` table columns ----
