+++ dango/archive/types/Cargo.toml
@@ -12,6 +12,15 @@ version       = { workspace = true }
 [lib]
 path = "src/lib.rs"
 
+[features]
+# Derives `utoipa::ToSchema` on the wire types, for the crates that serve them
+# over HTTP and document them via OpenAPI (the httpd and the projections).
+# Clients (e.g. dango-sdk) leave it off so they never pull utoipa.
+utoipa = ["dep:utoipa"]
+
 [dependencies]
 anyhow           = { workspace = true }
 dango-primitives = { workspace = true }
+serde            = { workspace = true, features = ["derive"] }
+serde_json       = { workspace = true }
+utoipa           = { workspace = true, optional = true }
